+++ collect_app/src/main/java/org/odk/collect/android/widgets/ExDecimalWidget.java
@@ -44,10 +44,6 @@ public class ExDecimalWidget extends ExStringWidget {
 
     public ExDecimalWidget(Context context, FormEntryPrompt prompt) {
         super(context, prompt);
-
-        answerText.setInputType(InputType.TYPE_NUMBER_FLAG_DECIMAL);
-        answerText.setKeyListener(new DigitsKeyListener(true, true));
-        setUpDecimalInputFilter();
     }
 
     @Override
